xnet2wordcloud/requirements.txt
matplotlib==3.1.1
nltk==3.4.5
numpy==1.17.0
pandas==0.25.0
python-igraph==0.7.1.post6
wordcloud==1.5.0
